+++ src/common/interfaces/IGatewaySenders.sol
@@ -12,11 +12,11 @@ interface ICentrifugeChainId {
 /// @notice Interface for dispatch-only gateway
 interface IPoolMessageSender is ICentrifugeChainId {
     /// @notice Creates and send the message
-    function sendNotifyPool(uint32 chainId, PoolId poolId) external;
+    function sendNotifyPool(uint16 chainId, PoolId poolId) external;
 
     /// @notice Creates and send the message
     function sendNotifyShareClass(
-        uint32 chainId,
+        uint16 chainId,
         PoolId poolId,
         ShareClassId scId,
         string memory name,
@@ -68,7 +68,7 @@ interface IPoolMessageSender is ICentrifugeChainId {
 /// @notice Interface for dispatch-only gateway
 interface IVaultMessageSender is ICentrifugeChainId {
     /// @notice Creates and send the message
-    function sendTransferShares(uint32 chainId, uint64 poolId, bytes16 scId, bytes32 recipient, uint128 amount)
+    function sendTransferShares(uint16 chainId, uint64 poolId, bytes16 scId, bytes32 recipient, uint128 amount)
         external;
 
     /// @notice Creates and send the message
@@ -87,7 +87,7 @@ interface IVaultMessageSender is ICentrifugeChainId {
 
     /// @notice Creates and send the message
     function sendRegisterAsset(
-        uint32 chainId,
+        uint16 chainId,
         uint128 assetId,
         string memory name,
         string memory symbol,
