+++ src/vaults/PoolManager.sol
@@ -102,7 +102,7 @@ contract PoolManager is Auth, IPoolManager, IUpdateContract, IPoolManagerGateway
     function transferTrancheTokens(
         uint64 poolId,
         bytes16 trancheId,
-        uint32 destinationId,
+        uint16 destinationId,
         bytes32 recipient,
         uint128 amount
     ) external auth {
@@ -116,7 +116,7 @@ contract PoolManager is Auth, IPoolManager, IUpdateContract, IPoolManagerGateway
     }
 
     // @inheritdoc IPoolManagerGatewayHandler
-    function registerAsset(address asset, uint256 tokenId, uint32 destChainId)
+    function registerAsset(address asset, uint256 tokenId, uint16 destChainId)
         external
         auth
         returns (uint128 assetId)
